--- a/crates/common/Cargo.toml
+++ b/crates/common/Cargo.toml
@@ -77,6 +77,8 @@ sysinfo = "0.35.1"
 proptest = { workspace = true }
 tempfile = { workspace = true }
 criterion = {workspace = true}
+rand.workspace = true
+regex = "1.11.1"
 
 [build-dependencies]
 cbindgen = { workspace = true, optional = true }