@faysou 2025-05-14
It's impressive. When most things will be in rust it will be possible to optimise other things that become dominant in time.